Fix "compare" function in "RunDownBean.getTopLeaks". (#411)
Rationale
The Skyline Nightly Test Results page displays an error if there are a lot of leaks reported because the "Compare" method implementation in "getTopLeaks" does not return consistent results. This was because it was mistakenly comparing the wrong set of numbers (comparing number of bytes leaked in one test to handle leaks in the other test) such that Compare(a, b) was not necessarily the opposite of Compare(b, a).